feat(json-schema): add conditional schema support
Add partial support for if/then/else clauses
in schemas, as per draft-07 json-schema specification.

This is to support showing/hiding form fields based
on values in the model to support dynamic forms from
a json-schema specification.

Supports adding fields in then/else clauses, but does
not support modifying attributes (such as 'minimum' or
'minLength') of already-existing fields.

Additionally, adds support for trivial schemas '{}',
'true', and 'false', as these are sometimes used in
conditions and then/else schemas.

Includes a variety of test cases.